refactor: restructure code
